+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Use standard FOLIO UX for Delete Course (in action menu). Fixes UICR-201.
 * When deleting a course that has no cross-listed courses, also delete its underlying courselisting. This ensures orphan courselistings are not created. Fixes UICR-206.
 * With orphan courselistings no longer remaining, it is now always possible to delete terms. Fixes UICR-203.
+* Provide local translations to ControlledVocab. Refs UICR-163.
+* Courses app: Replace collapse pane icons. Refs UICR-174.
+* App Context menu | Add new option | Courses app search. Refs UICR-166.
+* When a course is deleted, a green success toast displays, confirming deletion. Fixes UICR-207.
 
 ## [6.0.3](https://github.com/folio-org/ui-courses/tree/v6.0.3) (2023-10-20)